# [CSGen] Don't merge type variables for literals in `computeFavoredTypeForExpr`