We don't have enough information in CSGen to know that merging literal type variables is correct, and it doesn't seem to have a significant performance impact anyway.