@hborla 2020-07-22
@swift-ci please smoke test compiler performance