@hborla 2020-07-22
@swift-ci please test compiler performance