@xedin 2020-07-23
@hborla Looks like this change resulted in a couple more solver scopes but I think that's expected because we'd try more type variables which means more scopes.