@hborla 2020-07-23
@swift-ci please smoke test