@hborla 2020-07-23
# Summary for master full

**Unexpected test results, excluded stats for RxCocoa, Backend**

**Regressions found (see below)**

- [Debug-batch](#debug-batch-4e70a2f1d84b04ac)
    - [debug-batch brief](#debug-batch-brief-4e70a2f1d84b04ac)
    - [debug-batch detailed](#debug-batch-detailed-4e70a2f1d84b04ac)
- [Release](#release-4e70a2f1d84b04ac)
    - [release brief](#release-brief-4e70a2f1d84b04ac)
    - [release detailed](#release-detailed-4e70a2f1d84b04ac)



<a name="debug-batch-4e70a2f1d84b04ac">

# Debug-batch

<a name="debug-batch-brief-4e70a2f1d84b04ac">

## debug-batch brief
<details>
<summary>Regressed (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Improved (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (3)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Frontend.NumInstructionsExecuted | 71,559,631,842,321 | 71,506,832,648,261 | -52,799,194,060 | -0.07%
LLVM.NumLLVMBytesOutput | 1,599,314,472 | 1,599,123,282 | -191,190 | -0.01%
time.swift-driver.wall | 4321.6s | 4301.3s | -20.3s | -0.47%
</details>

<a name="debug-batch-detailed-4e70a2f1d84b04ac">

## debug-batch detailed
<details>
<summary>Regressed (2)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Driver.NumDriverPipePolls | 110,023 | 111,527 | 1,504 | 1.37% :no_entry:
Driver.NumDriverPipeReads | 103,086 | 104,575 | 1,489 | 1.44% :no_entry:
</details>
<details>
<summary>Improved (3)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Sema.LookupConformanceInModuleRequest | 36,044,462 | 35,198,115 | -846,347 | -2.35% :white_check_mark:
Sema.NumConstraintsConsideredForEdgeContraction | 94,248,448 | 86,219,281 | -8,029,167 | -8.52% :white_check_mark:
Sema.NumLeafScopes | 17,751,732 | 17,455,154 | -296,578 | -1.67% :white_check_mark:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (240)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.ImportSetCacheHit | 1,601,626 | 1,601,527 | -99 | -0.01%
AST.ImportSetCacheMiss | 320,819 | 320,817 | -2 | -0.0%
AST.ImportSetFoldHit | 108,043 | 108,044 | 1 | 0.0%
AST.ImportSetFoldMiss | 212,775 | 212,772 | -3 | -0.0%
AST.ModuleShadowCacheHit | 3,000 | 3,000 | 0 | 0.0%
AST.ModuleShadowCacheMiss | 1,293 | 1,293 | 0 | 0.0%
AST.ModuleVisibilityCacheHit | 33,242 | 33,242 | 0 | 0.0%
AST.ModuleVisibilityCacheMiss | 7,616 | 7,616 | 0 | 0.0%
AST.NumASTBytesAllocated | 32,913,010,338 | 32,903,591,227 | -9,419,111 | -0.03%
AST.NumASTScopeLookups | 3,109,128 | 3,109,204 | 76 | 0.0%
AST.NumBraceStmtASTScopeExpansions | 543,283 | 543,283 | 0 | 0.0%
AST.NumBraceStmtASTScopes | 543,283 | 543,283 | 0 | 0.0%
AST.NumDecls | 122,048 | 122,048 | 0 | 0.0%
AST.NumDependencies | 214,322 | 214,358 | 36 | 0.02%
AST.NumIterableTypeBodyASTScopeExpansions | 204,938 | 204,939 | 1 | 0.0%
AST.NumIterableTypeBodyASTScopes | 241,581 | 241,581 | 0 | 0.0%
AST.NumLinkLibraries | 0 | 0 | 0 | 0.0%
AST.NumLoadedModules | 185,428 | 185,428 | 0 | 0.0%
AST.NumLocalTypeDecls | 254 | 254 | 0 | 0.0%
AST.NumModuleLookupClassMember | 7,008 | 7,008 | 0 | 0.0%
AST.NumModuleLookupValue | 43,435,713 | 43,434,482 | -1,231 | -0.0%
AST.NumObjCMethods | 24,768 | 24,768 | 0 | 0.0%
AST.NumOperators | 683 | 683 | 0 | 0.0%
AST.NumPrecedenceGroups | 122 | 122 | 0 | 0.0%
AST.NumReferencedDynamicNames | 188 | 188 | 0 | 0.0%
AST.NumReferencedMemberNames | 7,887,079 | 7,886,941 | -138 | -0.0%
AST.NumReferencedTopLevelNames | 846,740 | 846,740 | 0 | 0.0%
AST.NumSourceBuffers | 269,182 | 269,182 | 0 | 0.0%
AST.NumSourceLines | 4,198,007 | 4,198,007 | 0 | 0.0%
AST.NumSourceLinesPerSecond | 2,092,280 | 2,093,879 | 1,599 | 0.08%
AST.NumTotalClangImportedEntities | 1,634,227 | 1,632,757 | -1,470 | -0.09%
Driver.ChildrenMaxRSS | 221,764,470,784 | 221,932,861,440 | 168,390,656 | 0.08%
Driver.DriverDepCascadingDynamic | 0 | 0 | 0 | 0.0%
Driver.DriverDepCascadingExternal | 0 | 0 | 0 | 0.0%
Driver.DriverDepCascadingMember | 0 | 0 | 0 | 0.0%
Driver.DriverDepCascadingNominal | 0 | 0 | 0 | 0.0%
Driver.DriverDepCascadingTopLevel | 0 | 0 | 0 | 0.0%
Driver.DriverDepDynamic | 0 | 0 | 0 | 0.0%
Driver.DriverDepExternal | 0 | 0 | 0 | 0.0%
Driver.DriverDepMember | 0 | 0 | 0 | 0.0%
Driver.DriverDepNominal | 0 | 0 | 0 | 0.0%
Driver.DriverDepTopLevel | 0 | 0 | 0 | 0.0%
Driver.NumDriverJobsRun | 24,308 | 24,308 | 0 | 0.0%
Driver.NumDriverJobsSkipped | 0 | 0 | 0 | 0.0%
Driver.NumProcessFailures | 0 | 0 | 0 | 0.0%
Frontend.MaxMallocUsage | 990,025,297,416 | 990,269,738,848 | 244,441,432 | 0.02%
Frontend.NumInstructionsExecuted | 71,559,631,842,321 | 71,506,832,648,261 | -52,799,194,060 | -0.07%
Frontend.NumProcessFailures | 0 | 0 | 0 | 0.0%
IRGen.IRGenRequest | 23,262 | 23,262 | 0 | 0.0%
IRGen.OptimizedIRRequest | 0 | 0 | 0 | 0.0%
IRModule.NumGOTEntries | 193,950 | 193,950 | 0 | 0.0%
IRModule.NumIRAliases | 177,131 | 177,131 | 0 | 0.0%
IRModule.NumIRBasicBlocks | 5,947,361 | 5,947,124 | -237 | -0.0%
IRModule.NumIRComdatSymbols | 0 | 0 | 0 | 0.0%
IRModule.NumIRFunctions | 2,990,393 | 2,990,314 | -79 | -0.0%
IRModule.NumIRGlobals | 3,310,170 | 3,310,012 | -158 | -0.0%
IRModule.NumIRIFuncs | 0 | 0 | 0 | 0.0%
IRModule.NumIRInsts | 79,480,659 | 79,479,402 | -1,257 | -0.0%
IRModule.NumIRNamedMetaData | 117,910 | 117,910 | 0 | 0.0%
IRModule.NumIRValueSymbols | 5,753,208 | 5,752,971 | -237 | -0.0%
LLVM.NumLLVMBytesOutput | 1,599,314,472 | 1,599,123,282 | -191,190 | -0.01%
Parse.CodeCompletionSecondPassRequest | 0 | 0 | 0 | 0.0%
Parse.NumFunctionsParsed | 240,875 | 240,875 | 0 | 0.0%
Parse.NumIterableDeclContextParsed | 537,502 | 537,503 | 1 | 0.0%
Parse.ParseAbstractFunctionBodyRequest | 219,610 | 219,610 | 0 | 0.0%
Parse.ParseMembersRequest | 424,608 | 424,609 | 1 | 0.0%
Parse.ParseSourceFileRequest | 266,183 | 266,183 | 0 | 0.0%
SILGen.ASTLoweringRequest | 24,303 | 24,303 | 0 | 0.0%
SILGen.ParseSILModuleRequest | 0 | 0 | 0 | 0.0%
SILModule.NumSILGenDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILGenFunctions | 1,838,887 | 1,838,892 | 5 | 0.0%
SILModule.NumSILGenGlobalVariables | 46,241 | 46,241 | 0 | 0.0%
SILModule.NumSILGenVtables | 29,036 | 29,036 | 0 | 0.0%
SILModule.NumSILGenWitnessTables | 108,056 | 108,056 | 0 | 0.0%
SILModule.NumSILOptDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILOptFunctions | 2,186,848 | 2,189,144 | 2,296 | 0.1%
SILModule.NumSILOptGlobalVariables | 47,335 | 47,335 | 0 | 0.0%
SILModule.NumSILOptVtables | 29,090 | 29,090 | 0 | 0.0%
SILModule.NumSILOptWitnessTables | 139,346 | 139,403 | 57 | 0.04%
SILOptimizer.ExecuteSILPipelineRequest | 94,149 | 94,149 | 0 | 0.0%
Sema.AbstractGenericSignatureRequest | 32,635 | 32,635 | 0 | 0.0%
Sema.AccessLevelRequest | 10,206,073 | 10,196,535 | -9,538 | -0.09%
Sema.AnyObjectLookupRequest | 276 | 276 | 0 | 0.0%
Sema.AreAllStoredPropertiesDefaultInitableRequest | 20,517 | 20,517 | 0 | 0.0%
Sema.AttachedFunctionBuilderRequest | 302,710 | 302,710 | 0 | 0.0%
Sema.AttachedPropertyWrapperTypeRequest | 500,522 | 500,522 | 0 | 0.0%
Sema.AttachedPropertyWrappersRequest | 2,298,564 | 2,298,565 | 1 | 0.0%
Sema.CallerSideDefaultArgExprRequest | 58,801 | 58,801 | 0 | 0.0%
Sema.CheckInconsistentImplementationOnlyImportsRequest | 6,871 | 6,871 | 0 | 0.0%
Sema.CheckRedeclarationRequest | 864,035 | 864,035 | 0 | 0.0%
Sema.ClassAncestryFlagsRequest | 92,681 | 92,681 | 0 | 0.0%
Sema.ClosureHasExplicitResultRequest | 78,694 | 78,694 | 0 | 0.0%
Sema.CodeCompletionFileRequest | 0 | 0 | 0 | 0.0%
Sema.CollectOverriddenDeclsRequest | 6,068,339 | 6,053,466 | -14,873 | -0.25%
Sema.CompareDeclSpecializationRequest | 398,660 | 398,167 | -493 | -0.12%
Sema.CursorInfoRequest | 0 | 0 | 0 | 0.0%
Sema.CustomAttrNominalRequest | 528 | 528 | 0 | 0.0%
Sema.CustomAttrTypeRequest | 338 | 338 | 0 | 0.0%
Sema.DefaultAndMaxAccessLevelRequest | 48,203 | 48,203 | 0 | 0.0%
Sema.DefaultArgumentExprRequest | 35,228 | 35,228 | 0 | 0.0%
Sema.DefaultArgumentInitContextRequest | 210 | 210 | 0 | 0.0%
Sema.DefaultDefinitionTypeRequest | 5,291 | 5,291 | 0 | 0.0%
Sema.DefaultTypeRequest | 330,538 | 330,538 | 0 | 0.0%
Sema.DerivativeAttrOriginalDeclRequest | 0 | 0 | 0 | 0.0%
Sema.DifferentiableAttributeTypeCheckRequest | 0 | 0 | 0 | 0.0%
Sema.DirectLookupRequest | 29,841,540 | 29,841,143 | -397 | -0.0%
Sema.DirectOperatorLookupRequest | 1,586,869 | 1,586,792 | -77 | -0.0%
Sema.DirectPrecedenceGroupLookupRequest | 553,392 | 553,392 | 0 | 0.0%
Sema.DynamicallyReplacedDeclRequest | 914,137 | 914,142 | 5 | 0.0%
Sema.EmittedMembersRequest | 24,187 | 24,187 | 0 | 0.0%
Sema.EnumRawTypeRequest | 19,455 | 19,455 | 0 | 0.0%
Sema.EnumRawValuesRequest | 9,483 | 9,483 | 0 | 0.0%
Sema.ExistentialConformsToSelfRequest | 11,439 | 11,440 | 1 | 0.01%
Sema.ExistentialTypeSupportedRequest | 13,723 | 13,723 | 0 | 0.0%
Sema.ExpandASTScopeRequest | 5,726,922 | 5,726,929 | 7 | 0.0%
Sema.ExtendedNominalRequest | 430,684 | 430,684 | 0 | 0.0%
Sema.ExtendedTypeRequest | 68,795 | 68,795 | 0 | 0.0%
Sema.FragileFunctionKindRequest | 1,508,148 | 1,508,153 | 5 | 0.0%
Sema.FunctionBuilderTypeRequest | 302,334 | 302,334 | 0 | 0.0%
Sema.FunctionOperatorRequest | 46,137 | 46,137 | 0 | 0.0%
Sema.GenericParamListRequest | 7,498,511 | 7,494,685 | -3,826 | -0.05%
Sema.GenericSignatureRequest | 1,993,916 | 1,993,661 | -255 | -0.01%
Sema.GetDestructorRequest | 24,844 | 24,844 | 0 | 0.0%
Sema.HasCircularInheritanceRequest | 21,145 | 21,145 | 0 | 0.0%
Sema.HasCircularInheritedProtocolsRequest | 8,513 | 8,513 | 0 | 0.0%
Sema.HasCircularRawValueRequest | 7,662 | 7,662 | 0 | 0.0%
Sema.HasDefaultInitRequest | 46,065 | 46,065 | 0 | 0.0%
Sema.HasDynamicCallableAttributeRequest | 0 | 0 | 0 | 0.0%
Sema.HasDynamicMemberLookupAttributeRequest | 493,350 | 493,283 | -67 | -0.01%
Sema.HasImplementationOnlyImportsRequest | 266,183 | 266,183 | 0 | 0.0%
Sema.HasMemberwiseInitRequest | 15,573 | 15,573 | 0 | 0.0%
Sema.HasMissingDesignatedInitializersRequest | 23,327 | 23,327 | 0 | 0.0%
Sema.HasUserDefinedDesignatedInitRequest | 46,075 | 46,075 | 0 | 0.0%
Sema.InferredGenericSignatureRequest | 151,660 | 151,663 | 3 | 0.0%
Sema.InheritedDeclsReferencedRequest | 4,393,131 | 4,390,103 | -3,028 | -0.07%
Sema.InheritedProtocolsRequest | 433,732 | 433,553 | -179 | -0.04%
Sema.InheritedTypeRequest | 257,063 | 256,985 | -78 | -0.03%
Sema.InheritsSuperclassInitializersRequest | 25,707 | 25,707 | 0 | 0.0%
Sema.InitKindRequest | 83,073 | 83,073 | 0 | 0.0%
Sema.InterfaceTypeRequest | 11,178,895 | 11,169,308 | -9,587 | -0.09%
Sema.IsABICompatibleOverrideRequest | 121,482 | 121,482 | 0 | 0.0%
Sema.IsAccessorTransparentRequest | 276,571 | 276,571 | 0 | 0.0%
Sema.IsCallableNominalTypeRequest | 1,638 | 1,638 | 0 | 0.0%
Sema.IsDeclApplicableRequest | 0 | 0 | 0 | 0.0%
Sema.IsDynamicRequest | 1,470,265 | 1,470,270 | 5 | 0.0%
Sema.IsFinalRequest | 2,266,975 | 2,264,735 | -2,240 | -0.1%
Sema.IsGetterMutatingRequest | 370,994 | 370,994 | 0 | 0.0%
Sema.IsImplicitlyUnwrappedOptionalRequest | 2,051,014 | 2,050,032 | -982 | -0.05%
Sema.IsObjCRequest | 1,316,921 | 1,316,970 | 49 | 0.0%
Sema.IsSetterMutatingRequest | 317,981 | 317,981 | 0 | 0.0%
Sema.IsStaticRequest | 784,887 | 784,905 | 18 | 0.0%
Sema.LazyStoragePropertyRequest | 2,336 | 2,336 | 0 | 0.0%
Sema.LookupAllConformancesInContextRequest | 72,195 | 72,195 | 0 | 0.0%
Sema.LookupInModuleRequest | 5,264,463 | 5,262,017 | -2,446 | -0.05%
Sema.LookupInfixOperatorRequest | 64,071 | 64,071 | 0 | 0.0%
Sema.LookupPostfixOperatorRequest | 184 | 184 | 0 | 0.0%
Sema.LookupPrecedenceGroupRequest | 24,001 | 24,001 | 0 | 0.0%
Sema.LookupPrefixOperatorRequest | 696 | 696 | 0 | 0.0%
Sema.MangleLocalTypeDeclRequest | 508 | 508 | 0 | 0.0%
Sema.ModuleImplicitImportsRequest | 7,892 | 7,892 | 0 | 0.0%
Sema.ModuleQualifiedLookupRequest | 2,408,091 | 2,405,572 | -2,519 | -0.1%
Sema.NamedLazyMemberLoadSuccessCount | 17,707,603 | 17,703,332 | -4,271 | -0.02%
Sema.NamingPatternRequest | 178,025 | 178,026 | 1 | 0.0%
Sema.NeedsNewVTableEntryRequest | 599,310 | 599,310 | 0 | 0.0%
Sema.NumAccessorBodiesSynthesized | 166,605 | 166,605 | 0 | 0.0%
Sema.NumAccessorsSynthesized | 249,854 | 249,854 | 0 | 0.0%
Sema.NumConformancesDeserialized | 7,216,373 | 7,218,727 | 2,354 | 0.03%
Sema.NumConstraintScopes | 27,853,971 | 27,924,429 | 70,458 | 0.25%
Sema.NumCrossImportsChecked | 0 | 0 | 0 | 0.0%
Sema.NumCrossImportsFound | 0 | 0 | 0 | 0.0%
Sema.NumCyclicOneWayComponentsCollapsed | 346 | 346 | 0 | 0.0%
Sema.NumDeclsDeserialized | 56,853,125 | 56,823,591 | -29,534 | -0.05%
Sema.NumDeclsTypechecked | 1,248,302 | 1,248,302 | 0 | 0.0%
Sema.NumGenericSignatureBuilders | 930,013 | 928,129 | -1,884 | -0.2%
Sema.NumLazyIterableDeclContexts | 6,586,644 | 6,584,873 | -1,771 | -0.03%
Sema.NumLazyRequirementSignatures | 612,365 | 612,363 | -2 | -0.0%
Sema.NumLazyRequirementSignaturesLoaded | 429,849 | 429,768 | -81 | -0.02%
Sema.NumTypesDeserialized | 17,900,837 | 17,881,170 | -19,667 | -0.11%
Sema.NumUnloadedLazyIterableDeclContexts | 4,086,683 | 4,086,627 | -56 | -0.0%
Sema.OpaqueReadOwnershipRequest | 239,698 | 239,698 | 0 | 0.0%
Sema.OpaqueResultTypeRequest | 203 | 203 | 0 | 0.0%
Sema.OperatorPrecedenceGroupRequest | 784 | 784 | 0 | 0.0%
Sema.OverriddenDeclsRequest | 2,090,039 | 2,084,844 | -5,195 | -0.25%
Sema.ParamSpecifierRequest | 1,062,326 | 1,062,326 | 0 | 0.0%
Sema.PatternBindingEntryRequest | 397,493 | 397,494 | 1 | 0.0%
Sema.PatternTypeRequest | 458,440 | 458,441 | 1 | 0.0%
Sema.PreCheckFunctionBuilderRequest | 330 | 330 | 0 | 0.0%
Sema.PrimarySourceFilesRequest | 7,892 | 7,892 | 0 | 0.0%
Sema.PropertyWrapperBackingPropertyInfoRequest | 495,252 | 495,252 | 0 | 0.0%
Sema.PropertyWrapperBackingPropertyTypeRequest | 500,522 | 500,522 | 0 | 0.0%
Sema.PropertyWrapperLValuenessRequest | 264 | 264 | 0 | 0.0%
Sema.PropertyWrapperMutabilityRequest | 526,443 | 526,443 | 0 | 0.0%
Sema.PropertyWrapperTypeInfoRequest | 92 | 92 | 0 | 0.0%
Sema.ProtocolRequiresClassRequest | 31,842 | 31,794 | -48 | -0.15%
Sema.ProvideDefaultImplForRequest | 6,068,339 | 6,053,466 | -14,873 | -0.25%
Sema.QualifiedLookupRequest | 4,675,817 | 4,674,596 | -1,221 | -0.03%
Sema.RangeInfoRequest | 0 | 0 | 0 | 0.0%
Sema.RequirementRequest | 104,278 | 104,278 | 0 | 0.0%
Sema.RequirementSignatureRequest | 475,592 | 475,432 | -160 | -0.03%
Sema.RequiresOpaqueAccessorsRequest | 1,144,489 | 1,144,489 | 0 | 0.0%
Sema.RequiresOpaqueModifyCoroutineRequest | 225,608 | 225,608 | 0 | 0.0%
Sema.ResolveEffectiveMemberwiseInitRequest | 0 | 0 | 0 | 0.0%
Sema.ResolveImplicitMemberRequest | 352,693 | 352,760 | 67 | 0.02%
Sema.ResolveProtocolNameRequest | 0 | 0 | 0 | 0.0%
Sema.ResolveTypeEraserTypeRequest | 0 | 0 | 0 | 0.0%
Sema.ResolveTypeRequest | 2,988,797 | 2,988,812 | 15 | 0.0%
Sema.ResultTypeRequest | 561,065 | 561,066 | 1 | 0.0%
Sema.RootAndResultTypeOfKeypathDynamicMemberRequest | 0 | 0 | 0 | 0.0%
Sema.RootTypeOfKeypathDynamicMemberRequest | 0 | 0 | 0 | 0.0%
Sema.SPIGroupsRequest | 10,544,471 | 10,528,842 | -15,629 | -0.15%
Sema.ScopedImportLookupRequest | 996 | 996 | 0 | 0.0%
Sema.SelfAccessKindRequest | 538,913 | 538,927 | 14 | 0.0%
Sema.SelfBoundsFromWhereClauseRequest | 1,439,395 | 1,438,380 | -1,015 | -0.07%
Sema.SetterAccessLevelRequest | 123,497 | 123,497 | 0 | 0.0%
Sema.SimpleDidSetRequest | 2,029,238 | 2,028,874 | -364 | -0.02%
Sema.StorageImplInfoRequest | 1,248,371 | 1,248,371 | 0 | 0.0%
Sema.StoredPropertiesAndMissingMembersRequest | 28,037 | 28,037 | 0 | 0.0%
Sema.StoredPropertiesRequest | 262,048 | 262,124 | 76 | 0.03%
Sema.StructuralTypeRequest | 1,412 | 1,412 | 0 | 0.0%
Sema.SuperclassDeclRequest | 332,522 | 332,343 | -179 | -0.05%
Sema.SuperclassTypeRequest | 47,712 | 47,712 | 0 | 0.0%
Sema.SynthesizeAccessorRequest | 249,854 | 249,854 | 0 | 0.0%
Sema.SynthesizeDefaultInitRequest | 4,520 | 4,520 | 0 | 0.0%
Sema.SynthesizeMainFunctionRequest | 23,461 | 23,461 | 0 | 0.0%
Sema.SynthesizeMemberwiseInitRequest | 2,959 | 2,959 | 0 | 0.0%
Sema.TangentStoredPropertyRequest | 0 | 0 | 0 | 0.0%
Sema.TypeCheckFunctionBodyAtLocRequest | 0 | 0 | 0 | 0.0%
Sema.TypeCheckFunctionBodyRequest | 469,446 | 469,446 | 0 | 0.0%
Sema.TypeCheckSourceFileRequest | 23,582 | 23,582 | 0 | 0.0%
Sema.TypeDeclsFromWhereClauseRequest | 26,493 | 26,493 | 0 | 0.0%
Sema.TypeEraserHasViableInitRequest | 0 | 0 | 0 | 0.0%
Sema.TypeRelationCheckRequest | 0 | 0 | 0 | 0.0%
Sema.TypeWitnessRequest | 7,799 | 7,799 | 0 | 0.0%
Sema.USRGenerationRequest | 7,184,246 | 7,169,801 | -14,445 | -0.2%
Sema.UnderlyingTypeDeclsReferencedRequest | 240,469 | 240,513 | 44 | 0.02%
Sema.UnderlyingTypeRequest | 30,593 | 30,593 | 0 | 0.0%
Sema.UnqualifiedLookupRequest | 3,234,278 | 3,234,354 | 76 | 0.0%
Sema.ValidatePrecedenceGroupRequest | 158,516 | 158,516 | 0 | 0.0%
Sema.ValueWitnessRequest | 56,279 | 56,279 | 0 | 0.0%
TBDGen.GenerateTBDRequest | 23,282 | 23,282 | 0 | 0.0%
</details>

<a name="release-4e70a2f1d84b04ac">

# Release

<a name="release-brief-4e70a2f1d84b04ac">

## release brief
<details>
<summary>Regressed (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Improved (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (3)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Frontend.NumInstructionsExecuted | 50,345,936,507,319 | 50,287,285,598,927 | -58,650,908,392 | -0.12%
LLVM.NumLLVMBytesOutput | 1,570,950,300 | 1,570,944,972 | -5,328 | -0.0%
time.swift-driver.wall | 5982.6s | 5966.4s | -16.2s | -0.27%
</details>

<a name="release-detailed-4e70a2f1d84b04ac">

## release detailed
<details>
<summary>Regressed (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Improved (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (16)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.NumLoadedModules | 21,335 | 21,335 | 0 | 0.0%
AST.NumTotalClangImportedEntities | 431,387 | 431,387 | 0 | 0.0%
IRModule.NumIRBasicBlocks | 4,005,561 | 4,005,561 | 0 | 0.0%
IRModule.NumIRFunctions | 2,451,779 | 2,451,779 | 0 | 0.0%
IRModule.NumIRGlobals | 2,955,117 | 2,955,117 | 0 | 0.0%
IRModule.NumIRInsts | 42,131,567 | 42,131,567 | 0 | 0.0%
IRModule.NumIRValueSymbols | 5,132,774 | 5,132,774 | 0 | 0.0%
LLVM.NumLLVMBytesOutput | 1,570,950,300 | 1,570,944,972 | -5,328 | -0.0%
SILModule.NumSILGenFunctions | 1,044,955 | 1,044,955 | 0 | 0.0%
SILModule.NumSILOptFunctions | 884,562 | 884,562 | 0 | 0.0%
Sema.NumConformancesDeserialized | 2,715,945 | 2,712,252 | -3,693 | -0.14%
Sema.NumConstraintScopes | 28,325,968 | 28,405,628 | 79,660 | 0.28%
Sema.NumDeclsDeserialized | 9,061,581 | 9,061,204 | -377 | -0.0%
Sema.NumGenericSignatureBuilders | 198,429 | 198,392 | -37 | -0.02%
Sema.NumLazyIterableDeclContexts | 1,147,355 | 1,147,407 | 52 | 0.0%
Sema.NumTypesDeserialized | 4,346,171 | 4,344,148 | -2,023 | -0.05%
</details>